+++ frontend/src/public/components/panels/summaryPanel.vue
@@ -1,14 +1,20 @@
 <template>
 	<div style="background-color: white">
-	<Toolbar :frame="frame" :toolbarStates="toolbarStates" />
-	<div v-if="clipping.Region.Summary" v-show="!clipping.Region.Summary.Empty" class="panel card panel-body"
-			style="background-color: transparent; padding-bottom: 13px">
-		<Clipping :clipping="clipping" :frame="frame"/>
-		<draggable v-model="propMetrics"  @end="itemMoved">
+		<Toolbar :frame="frame" :toolbarStates="toolbarStates" />
+		<div v-if="clipping.Region.Summary" v-show="!clipping.Region.Summary.Empty" class="panel card panel-body"
+				 style="background-color: transparent; padding-bottom: 13px">
+			<Clipping :clipping="clipping" :frame="frame" />
+			<draggable v-model="propMetrics" @end="itemMoved">
 				<transition-group name="fade">
 					<Metric v-for="(value, index) in metrics" :metric="value" :clipping="clipping" :key="index"></Metric>
 				</transition-group>
 			</draggable>
+		</div>
+		<div>
+			<WorkMetadata ref="showFuente" />
+			<MetricDownload ref="showDescargar" />
+			<MetricCustomize ref="showCustomize" />
+			<AddMetric ref="addMetric" />
 		</div>
 	</div>
 </template>
@@ -16,7 +22,11 @@
 <script>
 import Metric from '@/public/components/widgets/metric';
 import Clipping from '@/public/components/widgets/clipping';
+import WorkMetadata from '@/public/components/popups/workMetadata';
+import MetricCustomize from '@/public/components/popups/metricCustomize';
+import AddMetric from '@/public/components/popups/addMetric';
 import Toolbar from '@/public/components/widgets/toolbar';
+import MetricDownload from '@/public/components/popups/metricDownload';
 import draggable from 'vuedraggable';
 import arr from '@/common/js/arr';
 
@@ -25,6 +35,10 @@ export default {
 	components: {
 		Metric,
 		Clipping,
+		AddMetric,
+		WorkMetadata,
+		MetricCustomize,
+		MetricDownload,
 		Toolbar,
 		draggable
 	},
@@ -40,6 +54,12 @@ export default {
 			so: null,
 			back: null
 		};
+	},
+	mounted() {
+		window.Popups.MetricDownload = this.$refs.showDescargar;
+		window.Popups.WorkMetadata = this.$refs.showFuente;
+		window.Popups.MetricCustomize = this.$refs.showCustomize;
+		window.Popups.AddMetric = this.$refs.addMetric;
 	},
 	methods: {
 		itemMoved(evt) {