+++ frontend/src/public/googleMaps/GoogleMapsApi.js
@@ -103,7 +103,7 @@ GoogleMapsApi.prototype.Initialize = function () {
 	this.gMap = new this.google.maps.Map(document.getElementById('map'), myMapOptions);
 
 	this.gMap.mapTypes.set('blank', styledMapType);
-
+	this.AddCopyright();
 	this.CreateDrawingManager();
 	this.drawingManager.setMap(this.gMap);
 
@@ -141,6 +141,18 @@ GoogleMapsApi.prototype.Initialize = function () {
 		loc.CircleCompleted(circle);
 	});
 };
+
+GoogleMapsApi.prototype.AddCopyright = function () {
+	var controlDiv = document.createElement('DIV');
+	controlDiv.innerHTML = "<div class='copyrightText'>Poblaciones © 2019 CONICET/ODSA-UCA. " +
+		"<a class='copyrightText' href='https://poblaciones.org/terminos/' target='_blank'>Términos y Condiciones</a>. " +
+		"<a class='copyrightText' title='Comentarios y sugerencias a Poblaciones' href='https://poblaciones.org/contacto/' target='_blank'><i class='far fa-comments'></i> Contacto</a></div>";
+	controlDiv.className = "copyright";
+	controlDiv.index = 0;
+	this.gMap.controls[this.google.maps.ControlPosition.BOTTOM_RIGHT].push(controlDiv);
+
+};
+
 
 GoogleMapsApi.prototype.StopDrawing = function () {
 	this.drawingManager.setDrawingMode(null);